requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
scikit-learn
dill
pyaml
mlflow
xgboost
dagshub
fastapi
unicorn
-e .
